refactor(cli): isolate exec lifecycle and failure contracts
Split the exec owner into focused private modules, defer structured terminal publication until settlement and Patch delivery complete, and add deterministic provider, cancellation, resize, and process-timeout contracts.
