+++ .github/workflows/ci.yml
@@ -25,6 +25,7 @@ jobs:
   cli-test:
     name: CLI Tests (ubuntu-latest)
     runs-on: ubuntu-latest
+    timeout-minutes: 15
     steps:
       - uses: actions/checkout@v5
 
@@ -138,7 +139,7 @@ jobs:
 
       - name: Run Windows ConPTY CLI smoke test
         if: runner.os == 'Windows'
-        run: cargo test --locked -p bitfun-cli --test tui_terminal_process
+        run: cargo test --locked -p bitfun-cli --test terminal_process_contracts
 
       - name: Run core and desktop Rust tests
         run: cargo test --locked -p bitfun-core -p bitfun-desktop
